Refactor catalogInfoGenerator.ts and split under different files
