+++ plugins/bulk-import-backend/src/helpers/index.ts
@@ -15,7 +15,6 @@
  */
 
 export * from './auth';
-export * from './catalogInfoGenerator';
 export * from './GithubAppManager';
 export * from './loggingUtils';
 export * from './pagination';
