+++ plugins/bulk-import-backend/src/service/handlers/importStatus.ts
@@ -14,19 +14,21 @@
  * limitations under the License.
  */
 
-import type { AuthService, LoggerService } from '@backstage/backend-plugin-api';
-import type { CatalogApi } from '@backstage/catalog-client';
+import type { LoggerService } from '@backstage/backend-plugin-api';
 import type { Config } from '@backstage/config';
 
-import { getTokenForPlugin, type CatalogInfoGenerator } from '../../helpers';
 import type { Components } from '../../generated/openapi.d';
 import type { GithubApiService } from '../githubApiService';
+import { getCatalogUrl } from '../../catalog/catalogUtils';
+import {CatalogHttpClient} from "../../catalog/catalogHttpClient";
 
 export async function getImportStatusFromLocations(
-  logger: LoggerService,
-  config: Config,
-  githubApiService: GithubApiService,
-  catalogInfoGenerator: CatalogInfoGenerator,
+    deps: {
+        logger: LoggerService,
+        config: Config,
+        githubApiService: GithubApiService,
+        catalogHttpClient: CatalogHttpClient,
+    },
   repoUrl: string,
   catalogUrlLocations: string[],
   defaultBranch?: string,
@@ -35,10 +37,7 @@ export async function getImportStatusFromLocations(
   lastUpdate?: string;
 } | null> {
   return getImportStatusWithCheckerFn(
-    logger,
-    config,
-    githubApiService,
-    catalogInfoGenerator,
+    deps,
     repoUrl,
     async (catalogUrl: string) => {
       for (const loc of catalogUrlLocations) {
@@ -53,10 +52,12 @@ export async function getImportStatusFromLocations(
 }
 
 async function getImportStatusWithCheckerFn(
-  logger: LoggerService,
-  config: Config,
-  githubApiService: GithubApiService,
-  catalogInfoGenerator: CatalogInfoGenerator,
+  deps: {
+      logger: LoggerService,
+      config: Config,
+      githubApiService: GithubApiService,
+      catalogHttpClient: CatalogHttpClient,
+  },
   repoUrl: string,
   catalogExistenceCheckFn: (catalogUrl: string) => Promise<boolean>,
   defaultBranch?: string,
@@ -65,22 +66,21 @@ async function getImportStatusWithCheckerFn(
   lastUpdate?: string;
 } | null> {
   // Check to see if there are any PR
-  const openImportPr = await githubApiService.findImportOpenPr(logger, {
+  const openImportPr = await deps.githubApiService.findImportOpenPr(deps.logger, {
     repoUrl,
   });
   if (!openImportPr.prUrl) {
     const existsInCatalog = await catalogExistenceCheckFn(
-      catalogInfoGenerator.getCatalogUrl(config, repoUrl, defaultBranch),
+      getCatalogUrl(deps.config, repoUrl, defaultBranch),
     );
     const existsInRepo =
-      await githubApiService.doesCatalogInfoAlreadyExistInRepo(logger, {
+      await deps.githubApiService.doesCatalogInfoAlreadyExistInRepo(deps.logger, {
         repoUrl,
         defaultBranch,
       });
     if (existsInCatalog && existsInRepo) {
       // Force a refresh of the Location, so that the entities from the catalog-info.yaml can show up quickly (not guaranteed however).
-      await catalogInfoGenerator.refreshLocationByRepoUrl(
-        config,
+      await deps.catalogHttpClient.refreshLocationByRepoUrl(
         repoUrl,
         defaultBranch,
       );
@@ -89,58 +89,4 @@ async function getImportStatusWithCheckerFn(
     return null;
   }
   return { status: 'WAIT_PR_APPROVAL', lastUpdate: openImportPr.lastUpdate };
-}
-
-/**
- * verifyLocationExistence checks for the existence of the Location target.
- * Under the hood, it attempts to read the target URL and will return false if the target could not be found
- * and even if there is already a Location row in the database.
- * @param auth
- * @param catalogApi
- * @param repoCatalogUrl
- */
-export async function verifyLocationExistence(
-  auth: AuthService,
-  catalogApi: CatalogApi,
-  repoCatalogUrl: string,
-): Promise<boolean> {
-  try {
-    const result = await catalogApi.addLocation(
-      {
-        type: 'url',
-        target: repoCatalogUrl,
-        dryRun: true,
-      },
-      {
-        token: await getTokenForPlugin(auth, 'catalog'),
-      },
-    );
-    // The `result.exists` field is only filled in dryRun mode
-    return result.exists as boolean;
-  } catch (error: any) {
-    if (error.message?.includes('NotFoundError')) {
-      return false;
-    }
-    throw error;
-  }
-}
-
-export async function hasEntityInCatalog(
-  auth: AuthService,
-  catalogApi: CatalogApi,
-  entityName: string,
-) {
-  return catalogApi
-    .queryEntities(
-      {
-        filter: {
-          'metadata.name': entityName,
-        },
-        limit: 1,
-      },
-      {
-        token: await getTokenForPlugin(auth, 'catalog'),
-      },
-    )
-    .then(resp => resp.items?.length > 0);
 }
